+++ essencium-backend-sequence-model/src/main/java/de/frachtwerk/essencium/backend/model/User.java
@@ -39,4 +39,9 @@ public class User extends AbstractBaseUser<Long> {
       sequenceName = "hibernate_sequence",
       allocationSize = 1)
   private Long id;
+
+  @Override
+  public String getTitle() {
+    return getFirstName() + " " + getLastName();
+  }
 }