+++ essencium-backend/src/main/java/de/frachtwerk/essencium/backend/controller/AbstractAccessAwareController.java
@@ -23,6 +23,7 @@ import de.frachtwerk.essencium.backend.controller.access.RestrictAccessToOwnedEn
 import de.frachtwerk.essencium.backend.model.AbstractBaseModel;
 import de.frachtwerk.essencium.backend.model.Identifiable;
 import de.frachtwerk.essencium.backend.model.exception.ResourceNotFoundException;
+import de.frachtwerk.essencium.backend.model.representation.BasicRepresentation;
 import de.frachtwerk.essencium.backend.service.AbstractEntityService;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
@@ -32,6 +33,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
 import java.io.Serializable;
+import java.util.List;
 import java.util.Map;
 import java.util.Set;
 import lombok.AllArgsConstructor;
@@ -142,6 +144,61 @@ public abstract class AbstractAccessAwareController<
   public Page<REPRESENTATION> findAll(
       @Parameter(hidden = true) SPEC specification, @ParameterObject Pageable pageable) {
     return toRepresentation(service.getAllFiltered(specification, pageable));
+  }
+
+  @GetMapping("/basic")
+  @Parameter(
+      in = ParameterIn.QUERY,
+      name = "ids",
+      description =
+          "IDs of the requested entities. can contain multiple values separated by ','"
+              + "Multiple criteria are supported.",
+      content =
+          @Content(array = @ArraySchema(schema = @Schema(type = "integer", example = "1,2,5"))))
+  @Parameter(
+      in = ParameterIn.QUERY,
+      name = "createdBy",
+      description = "full username (email)",
+      content = @Content(schema = @Schema(type = "string", example = "devnull@frachtwerk.de")))
+  @Parameter(
+      in = ParameterIn.QUERY,
+      name = "updatedBy",
+      description = "full username (email)",
+      content = @Content(schema = @Schema(type = "string", example = "devnull@frachtwerk.de")))
+  @Parameter(
+      in = ParameterIn.QUERY,
+      name = "createdAtFrom",
+      description = "returns entries created after the submitted date and time ",
+      content =
+          @Content(
+              schema =
+                  @Schema(type = "string", format = "date-time", example = "2021-01-01T00:00:01")))
+  @Parameter(
+      in = ParameterIn.QUERY,
+      name = "createdAtTo",
+      description = "returns entries created before the submitted date and time ",
+      content =
+          @Content(
+              schema =
+                  @Schema(type = "string", format = "date-time", example = "2021-12-31T23:59:59")))
+  @Parameter(
+      in = ParameterIn.QUERY,
+      name = "updatedAtFrom",
+      description = "returns entries updated after the submitted date and time ",
+      content =
+          @Content(
+              schema =
+                  @Schema(type = "string", format = "date-time", example = "2021-01-01T00:00:01")))
+  @Parameter(
+      in = ParameterIn.QUERY,
+      name = "updatedAtTo",
+      description = "returns entries updated before the submitted date and time ",
+      content =
+          @Content(
+              schema =
+                  @Schema(type = "string", format = "date-time", example = "2021-12-31T23:59:59")))
+  public List<BasicRepresentation> findAll(@Parameter(hidden = true) SPEC specification) {
+    return BasicRepresentation.from(service.getAllFiltered(specification));
   }
 
   @GetMapping("/{id}")