+++ essencium-backend/src/main/java/de/frachtwerk/essencium/backend/model/TitleConvention.java
@@ -1,0 +1,9 @@
+package de.frachtwerk.essencium.backend.model;
+
+import java.io.Serializable;
+
+public interface TitleConvention<ID extends Serializable> {
+  String getTitle();
+
+  ID getId();
+}