+++ essencium-backend/src/test/java/de/frachtwerk/essencium/backend/service/AbstractCrudServiceTest.java
@@ -324,6 +324,11 @@ class AbstractCrudServiceTest {
   @AllArgsConstructor
   static class TestSequenceIdModel extends SequenceIdModel {
     private String identifier;
+
+    @Override
+    public String getTitle() {
+      return identifier;
+    }
   }
 
   interface TestRepo extends BaseRepository<TestSequenceIdModel, Long> {}