+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version `______`
 
+- Feature: `BasicRepresentation` can be used to create slim representations of entities. `BasicRepresentation` is very well suited for filling embedded objects or lists of objects in other representations.
+- Feature: Each Controller now has a `GET .../basic` endpoint that returns a List of BasicRepresentations of the entity. Filtering via Specifications is possible
 - upgrade io.sentry:sentry-spring-boot-starter-jakarta from 7.12.0 to 7.13.0
 - upgraded org.hibernate.orm:hibernate-jpamodelgen from 6.5.2.Final to 6.6.0.Final